add auth mutations
