+++ frontend/src/APIClients/mutations/AuthMutations.ts
@@ -1,0 +1,59 @@
+import { gql } from "@apollo/client";
+
+export const LOGIN = gql`
+  mutation Login($email: String!, $password: String!) {
+    login(email: $email, password: $password) {
+      id
+      firstName
+      lastName
+      email
+      role
+      approvedLanguagesTranslation
+      approvedLanguagesReview
+      accessToken
+      refreshToken
+    }
+  }
+`;
+
+export const REFRESH = gql`
+  mutation Refresh {
+    refresh
+  }
+`;
+
+export const RESET_PASSWORD = gql`
+  mutation ResetPassword($email: String!) {
+    resetPassword(email: $email) {
+      ok
+    }
+  }
+`;
+
+export const REGISTER = gql`
+  mutation Signup_Register(
+    $firstName: String!
+    $lastName: String!
+    $email: String!
+    $town: String!
+    $password: String!
+  ) {
+    register(
+      user: {
+        firstName: $firstName
+        lastName: $lastName
+        email: $email
+        town: $town
+        password: $password
+      }
+    ) {
+      id
+      firstName
+      lastName
+      email
+      town
+      role
+      accessToken
+    }
+  }
+`;
